requirements/dev.txt
-r base.txt

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Development
ipython==8.19.0